Add JWT handler for AWS Cognito
This adds a handler for `aws` when `JWT_MODULE` is set to `aws`. The handler logic
then looks for `JWK_URL` to be set, where the JWK set will be retrieved.

In AWS Cognito, this consists of a full JWK set of keys. We need to match the
key identifier (`kid`) from a token to determine which JWK we need to build a PEM
pub key from. Once we know this, we'll use `jwk2pem` [1] to match the `kid` with
the correct JWK and convert the key to PEM to return in the payload.

TODO:
- Cache JWK response
- Adjust the payload to return the PEM data with a `pubkey` key in a JSON object,
  like: `{"pubkey": "PEM DATA"}`

[1] https://github.com/RedHatInsights/jwk2pem