Revert back to pubkey return for ephemeral
Reverts some of the changes from 998356587be35eaf2b2f8208b5b4775755be5fdb in order
to not break the ephemeral API contract for the Caddy sidecar.

Long term it would be good to modify the sidecar to expect a `pubkey` key on the
JSON response [1], in order to have parity across providers on the `/v1/jwt` endpoint.

[1] https://github.com/RedHatInsights/crcauthlib/blob/master/crcauthlib.go#L73