Add token signing and handler
Creates a token model which can create a JWT provided a TTL and identity object
from an `x-rh-identity` header.

The `Create` method will parse the private key, build the following claims:

```
exp - expiration of the token
iat - time the token was issued
nbf - time the token is not good before
kid - identifier for JWK (may not be required)
org_id - from the identity
username - from the identity
```

Those claims are then signed with the private key. We may not explicitly need
to use the public key for this, but we'll either need to surface the public key
for the gateway to consume via an endpoint, or secret.

This also sets up the handler, which will service `/v1/registrations/token` GET
requests.

We'll enforce the `x-rh-identity` exists, as well as `org_id` and `username` as
we'll need those for the token.

We create a new token object with the pub/private keys, set the TTL, and pass the
identity and TTL into the `Create` method, handling errors before returning a
token payload like:

```
{"token": <token>}
```